+++ CHANGELOG.md
@@ -10,6 +10,7 @@ for a given releases. Unreleased, upcoming changes will be updated here periodic
 
 - Fix bug that we did not delete the webp variant when enabled, when deleting an image through the CacheManager ([ousamabenyounes](https://github.com/liip/LiipImagineBundle/pull/1661))
 - Drop support for PHP 8.0.
+- Throw `InvalidSignedUrlException` (extends `BadRequestHttpException`) when a runtime signed URL fails the HMAC check ([BriceFab](https://github.com/liip/LiipImagineBundle/issues/1656))
 
 ## [2.17.2](https://github.com/liip/LiipImagineBundle/tree/2.17.2)